requirements.txt
bedrock-agentcore-starter-toolkit==0.2.8
boto3==1.42.36
botocore==1.42.36
duckduckgo-search==8.1.1
ipykernel==7.1.0
ipython==9.9.0
jupyter==1.1.1
langfuse==3.12.1
mcp==1.26.0
numpy==2.4.1
pandas==3.0.0
pydantic==2.12.5
pytest-asyncio==1.3.0
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.1
ruff==0.14.14
scikit-learn==1.8.0
strands-agents-tools==0.2.19
strands-agents[otel]==1.23.0
tabulate==0.9.0